fix: move `CalculatorDeviceRow` to its own file
