+++ src/app/energyCalculator/components/CalculatorDeviceRow.tsx
